Implementing US 5182: removed container configuration stuff and added parameter for long term session life time

+++ sessiond/src/com/openexchange/sessiond/impl/SessionContainer.java
@@ -164,7 +164,7 @@ final class SessionContainer {
      * @param timeToLive The session's time to live
      * @return The wrapping {@link SessionControl session control}.
      */
-    SessionControl put(final Session session, final int timeToLive) {
+    SessionControl put(final Session session, final long timeToLive) {
         final String sessionId = session.getSessionID();
         /*
          * Add session
